+++ cvat-canvas/src/typescript/canvasModel.ts
@@ -46,6 +46,11 @@ export enum RectDrawingMethod {
     EXTREME_POINTS = 'By 4 points'
 }
 
+export interface Configuration {
+    displayAllText?: boolean;
+    undefinedAttrValue?: string;
+}
+
 export interface DrawData {
     enabled: boolean;
     shapeType?: string;
@@ -101,6 +106,7 @@ export enum UpdateReasons {
     BITMAP = 'bitmap',
     DRAG_CANVAS = 'drag_canvas',
     ZOOM_CANVAS = 'zoom_canvas',
+    CONFIG_UPDATED = 'config_updated',
 }
 
 export enum Mode {
@@ -128,6 +134,7 @@ export interface CanvasModel {
     readonly mergeData: MergeData;
     readonly splitData: SplitData;
     readonly groupData: GroupData;
+    readonly configuration: Configuration;
     readonly selected: any;
     geometry: Geometry;
     mode: Mode;
@@ -154,6 +161,7 @@ export interface CanvasModel {
     dragCanvas(enable: boolean): void;
     zoomCanvas(enable: boolean): void;
 
+    configure(configuration: Configuration): void;
     cancel(): void;
 }
 
@@ -162,6 +170,7 @@ export class CanvasModelImpl extends MasterImpl implements CanvasModel {
         activeElement: ActiveElement;
         angle: number;
         canvasSize: Size;
+        configuration: Configuration;
         imageBitmap: boolean;
         image: Image | null;
         imageID: number | null;
@@ -194,6 +203,10 @@ export class CanvasModelImpl extends MasterImpl implements CanvasModel {
             canvasSize: {
                 height: 0,
                 width: 0,
+            },
+            configuration: {
+                displayAllText: false,
+                undefinedAttrValue: '',
             },
             imageBitmap: false,
             image: null,
@@ -495,8 +508,28 @@ export class CanvasModelImpl extends MasterImpl implements CanvasModel {
         this.data.selected = null;
     }
 
+    public configure(configuration: Configuration): void {
+        if (this.data.mode !== Mode.IDLE) {
+            throw Error(`Canvas is busy. Action: ${this.data.mode}`);
+        }
+
+        if (typeof (configuration.displayAllText) !== 'undefined') {
+            this.data.configuration.displayAllText = configuration.displayAllText;
+        }
+
+        if (typeof (configuration.undefinedAttrValue) !== 'undefined') {
+            this.data.configuration.undefinedAttrValue = configuration.undefinedAttrValue;
+        }
+
+        this.notify(UpdateReasons.CONFIG_UPDATED);
+    }
+
     public cancel(): void {
         this.notify(UpdateReasons.CANCEL);
+    }
+
+    public get configuration(): Configuration {
+        return { ...this.data.configuration };
     }
 
     public get geometry(): Geometry {